Handle the export space cases
