+++ .gitignore
@@ -2,3 +2,4 @@ tester.sh
 .vscode/
 .nfs0000000001380ae200000017
 .gitignore
+.nfs000000000138013400000001
