+++ Makefile
@@ -6,7 +6,7 @@
 #    By: malrifai <malrifai@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/02/03 10:28:41 by eaqrabaw          #+#    #+#              #
-#    Updated: 2025/02/22 15:57:18 by malrifai         ###   ########.fr        #
+#    Updated: 2025/02/22 19:19:35 by malrifai         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -28,7 +28,7 @@ EXPANDER	= expander/expand.c expander/utils.c expander/var_expand.c
 
 PARSING		=
 
-BUILTINS	= builtins/echo.c builtins/export.c builtins/pwd.c builtins/env.c builtins/env_utils.c
+BUILTINS	= builtins/echo.c builtins/export/export.c builtins/export/print_env.c builtins/export/check.c builtins/pwd.c builtins/env.c builtins/env_utils.c
 
 SIGNAL		= signal/signal_handler.c
